EhsEnforcement.Agencies.Ea.OffenderBuilder (ehs_enforcement v0.1.0)

Builds offender attributes for EA (Environment Agency) enforcement data.

This module consolidates offender attribute building logic that was previously duplicated across EA.CaseProcessor and EA.NoticeProcessor.

Usage

For case data (EaDetailRecord):

iex> ea_record = %EaDetailRecord{offender_name: "ABC Water Treatment Ltd", ...}
iex> OffenderBuilder.build_offender_attrs(ea_record, :case)
%{name: "ABC Water Treatment Ltd", business_type: :limited_company, ...}

For notice data (map):

iex> notice_data = %{offender_name: "XYZ Environmental Services", ...}
iex> OffenderBuilder.build_offender_attrs(notice_data, :notice)
%{offender_name: "XYZ Environmental Services", ...}

Data Types

:case - EA case data from enforcement action scraping

:notice - EA notice data from notice scraping

EA-Specific Fields

EA offenders include unique fields not present in HSE data:

:company_registration_number - Company registration number

:town - Town (separate from address)

:county - County

:postcode - Postcode

:industry_sector - EA industry classification

Functions

build_offender_attrs(data, atom)

Builds offender attributes from EA case or notice data.

Parameters

data - Either an %EaDetailRecord{} struct (for cases) or a map (for notices)

data_type - Either :case or :notice to indicate the data source

Returns

A map of offender attributes with:

Common fields: :name, :address, :business_type

EA case fields: :local_authority, :postcode, :main_activity, :industry, :company_registration_number, :town, :county

EA notice fields: :offender_name, :offender_address, :company_registration_number, :industry_sector

Nil and empty string values are automatically filtered out.

Examples

# Case data
ea_record = %EaDetailRecord{
  offender_name: "ABC Water Ltd",
  county: "Yorkshire",
  industry_sector: "Water Supply"
}
OffenderBuilder.build_offender_attrs(ea_record, :case)
# => %{
#   name: "ABC Water Ltd",
#   business_type: :limited_company,
#   local_authority: "Yorkshire",
#   industry: "Extractive and utility supply industries"
# }

# Notice data
notice_data = %{
  offender_name: "XYZ Environmental Services",
  address: "123 Green Street",
  town: "Manchester",
  county: "Greater Manchester"
}
OffenderBuilder.build_offender_attrs(notice_data, :notice)
# => %{
#   offender_name: "XYZ Environmental Services",
#   offender_address: "123 Green Street, Manchester, Greater Manchester"
# }