EhsEnforcement.Enforcement.Offender (ehs_enforcement v0.1.0)

Represents a company or individual subject to enforcement action. Normalized to eliminate duplication between cases and notices.

Types

t()

@type t() :: %EhsEnforcement.Enforcement.Offender{
  __lateral_join_source__: term(),
  __meta__: term(),
  __metadata__: term(),
  __order__: term(),
  address: term(),
  agencies: term(),
  aggregates: term(),
  business_type: term(),
  calculations: term(),
  cases: term(),
  company_registration_number: term(),
  country: term(),
  county: term(),
  enforcement_count: term(),
  first_seen_date: term(),
  id: term(),
  industry: term(),
  industry_sectors: term(),
  inserted_at: term(),
  last_seen_date: term(),
  local_authority: term(),
  main_activity: term(),
  name: term(),
  normalized_name: term(),
  notices: term(),
  postcode: term(),
  sic_code: term(),
  total_cases: term(),
  total_fines: term(),
  total_notices: term(),
  town: term(),
  updated_at: term()
}

Functions

can_create(actor, name, params_or_opts \\ %{}, opts \\ [])

Runs authorization checks for EhsEnforcement.Enforcement.Offender.create See Ash.can/3 for more information

Options

:maybe_is (term/0) - If the actor may be able to perform the action, what value should be returned. The default value is :maybe.

:filter_with - If set to :error, the query will raise an error on a match. If set to :filter the query will filter out unauthorized access. Valid values are :filter, :error The default value is :filter.

:validate? (boolean/0) - Whether or not to treat an invalid action as a non-allowed action. The default value is false.

:reuse_values? (boolean/0) - Whether or not loaded data like aggregates, calculations and relationships should be checked in memory if possible, instead of querying. No effect if pre_flight? is false. The default value is false.

:pre_flight? (boolean/0) - Whether or not this is a pre_flight check (which may perform optimized in-memory checks) or the final proper check. The default value is true.

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol. Will overwrite any actor, tenant or context provided. See Ash.Context for more.

:run_queries? (boolean/0) - Whether or not to run queries. If set to true, :maybe will not be returned. The default value is true.

:data - The record or records specifically attempting to be acted upon.

:tenant (value that implements the Ash.ToTenant protocol) - The tenant to use for authorization

:alter_source? (boolean/0) - If set to true, the source being authorized is returned so it can be run. The default value is false.

:base_query (term/0) - A base query on which to apply an generated filters

:no_check? (boolean/0) - Whether or not authorization must pass at the strict/filter step, or if post-checks are allowed to be run The default value is false.

:on_must_pass_strict_check (term/0) - Override the value returned when no_check? is true but a check must be run.

:atomic_changeset (term/0) - A base query on which to apply an generated filters

:return_forbidden_error? (boolean/0) - Whether or not to return a forbidden error in cases of not being authorized. The default value is false.

:log? (boolean/0) - Whether or not to log the authorization result. The default value is false.

can_create?(actor, name, params_or_opts \\ %{}, opts \\ [])

Runs authorization checks for EhsEnforcement.Enforcement.Offender.create, returning a boolean. See Ash.can?/3 for more information

Options

:maybe_is (term/0) - If the actor may be able to perform the action, what value should be returned. The default value is true.

:filter_with - If set to :error, the query will raise an error on a match. If set to :filter the query will filter out unauthorized access. Valid values are :filter, :error The default value is :filter.

:validate? (boolean/0) - Whether or not to treat an invalid action as a non-allowed action. The default value is false.

:reuse_values? (boolean/0) - Whether or not loaded data like aggregates, calculations and relationships should be checked in memory if possible, instead of querying. No effect if pre_flight? is false. The default value is false.

:pre_flight? (boolean/0) - Whether or not this is a pre_flight check (which may perform optimized in-memory checks) or the final proper check. The default value is true.

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol. Will overwrite any actor, tenant or context provided. See Ash.Context for more.

:run_queries? (boolean/0) - Whether or not to run queries. If set to true, :maybe will not be returned. The default value is true.

:data - The record or records specifically attempting to be acted upon.

:tenant (value that implements the Ash.ToTenant protocol) - The tenant to use for authorization

:alter_source? (boolean/0) - If set to true, the source being authorized is returned so it can be run. The default value is false.

:base_query (term/0) - A base query on which to apply an generated filters

:no_check? (boolean/0) - Whether or not authorization must pass at the strict/filter step, or if post-checks are allowed to be run The default value is false.

:on_must_pass_strict_check (term/0) - Override the value returned when no_check? is true but a check must be run.

:atomic_changeset (term/0) - A base query on which to apply an generated filters

:return_forbidden_error? (boolean/0) - Whether or not to return a forbidden error in cases of not being authorized. The default value is false.

:log? (boolean/0) - Whether or not to log the authorization result. The default value is false.

can_search(actor, query, params_or_opts \\ %{}, opts \\ [])

Runs authorization checks for EhsEnforcement.Enforcement.Offender.search See Ash.can/3 for more information

Options

:maybe_is (term/0) - If the actor may be able to perform the action, what value should be returned. The default value is :maybe.

:filter_with - If set to :error, the query will raise an error on a match. If set to :filter the query will filter out unauthorized access. Valid values are :filter, :error The default value is :filter.

:validate? (boolean/0) - Whether or not to treat an invalid action as a non-allowed action. The default value is false.

:reuse_values? (boolean/0) - Whether or not loaded data like aggregates, calculations and relationships should be checked in memory if possible, instead of querying. No effect if pre_flight? is false. The default value is false.

:pre_flight? (boolean/0) - Whether or not this is a pre_flight check (which may perform optimized in-memory checks) or the final proper check. The default value is true.

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol. Will overwrite any actor, tenant or context provided. See Ash.Context for more.

:run_queries? (boolean/0) - Whether or not to run queries. If set to true, :maybe will not be returned. The default value is true.

:data - The record or records specifically attempting to be acted upon.

:tenant (value that implements the Ash.ToTenant protocol) - The tenant to use for authorization

:alter_source? (boolean/0) - If set to true, the source being authorized is returned so it can be run. The default value is false.

:base_query (term/0) - A base query on which to apply an generated filters

:no_check? (boolean/0) - Whether or not authorization must pass at the strict/filter step, or if post-checks are allowed to be run The default value is false.

:on_must_pass_strict_check (term/0) - Override the value returned when no_check? is true but a check must be run.

:atomic_changeset (term/0) - A base query on which to apply an generated filters

:return_forbidden_error? (boolean/0) - Whether or not to return a forbidden error in cases of not being authorized. The default value is false.

:log? (boolean/0) - Whether or not to log the authorization result. The default value is false.

can_search?(actor, query, params_or_opts \\ %{}, opts \\ [])

Runs authorization checks for EhsEnforcement.Enforcement.Offender.search, returning a boolean. See Ash.can?/3 for more information

Options

:maybe_is (term/0) - If the actor may be able to perform the action, what value should be returned. The default value is true.

:filter_with - If set to :error, the query will raise an error on a match. If set to :filter the query will filter out unauthorized access. Valid values are :filter, :error The default value is :filter.

:validate? (boolean/0) - Whether or not to treat an invalid action as a non-allowed action. The default value is false.

:reuse_values? (boolean/0) - Whether or not loaded data like aggregates, calculations and relationships should be checked in memory if possible, instead of querying. No effect if pre_flight? is false. The default value is false.

:pre_flight? (boolean/0) - Whether or not this is a pre_flight check (which may perform optimized in-memory checks) or the final proper check. The default value is true.

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol. Will overwrite any actor, tenant or context provided. See Ash.Context for more.

:run_queries? (boolean/0) - Whether or not to run queries. If set to true, :maybe will not be returned. The default value is true.

:data - The record or records specifically attempting to be acted upon.

:tenant (value that implements the Ash.ToTenant protocol) - The tenant to use for authorization

:alter_source? (boolean/0) - If set to true, the source being authorized is returned so it can be run. The default value is false.

:base_query (term/0) - A base query on which to apply an generated filters

:no_check? (boolean/0) - Whether or not authorization must pass at the strict/filter step, or if post-checks are allowed to be run The default value is false.

:on_must_pass_strict_check (term/0) - Override the value returned when no_check? is true but a check must be run.

:atomic_changeset (term/0) - A base query on which to apply an generated filters

:return_forbidden_error? (boolean/0) - Whether or not to return a forbidden error in cases of not being authorized. The default value is false.

:log? (boolean/0) - Whether or not to log the authorization result. The default value is false.

can_update_statistics(actor, record, params_or_opts \\ %{}, opts \\ [])

Runs authorization checks for EhsEnforcement.Enforcement.Offender.update_statistics See Ash.can/3 for more information

Options

:maybe_is (term/0) - If the actor may be able to perform the action, what value should be returned. The default value is :maybe.

:filter_with - If set to :error, the query will raise an error on a match. If set to :filter the query will filter out unauthorized access. Valid values are :filter, :error The default value is :filter.

:validate? (boolean/0) - Whether or not to treat an invalid action as a non-allowed action. The default value is false.

:reuse_values? (boolean/0) - Whether or not loaded data like aggregates, calculations and relationships should be checked in memory if possible, instead of querying. No effect if pre_flight? is false. The default value is false.

:pre_flight? (boolean/0) - Whether or not this is a pre_flight check (which may perform optimized in-memory checks) or the final proper check. The default value is true.

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol. Will overwrite any actor, tenant or context provided. See Ash.Context for more.

:run_queries? (boolean/0) - Whether or not to run queries. If set to true, :maybe will not be returned. The default value is true.

:data - The record or records specifically attempting to be acted upon.

:tenant (value that implements the Ash.ToTenant protocol) - The tenant to use for authorization

:alter_source? (boolean/0) - If set to true, the source being authorized is returned so it can be run. The default value is false.

:base_query (term/0) - A base query on which to apply an generated filters

:no_check? (boolean/0) - Whether or not authorization must pass at the strict/filter step, or if post-checks are allowed to be run The default value is false.

:on_must_pass_strict_check (term/0) - Override the value returned when no_check? is true but a check must be run.

:atomic_changeset (term/0) - A base query on which to apply an generated filters

:return_forbidden_error? (boolean/0) - Whether or not to return a forbidden error in cases of not being authorized. The default value is false.

:log? (boolean/0) - Whether or not to log the authorization result. The default value is false.

can_update_statistics?(actor, record, params_or_opts \\ %{}, opts \\ [])

Runs authorization checks for EhsEnforcement.Enforcement.Offender.update_statistics, returning a boolean. See Ash.can?/3 for more information

Options

:maybe_is (term/0) - If the actor may be able to perform the action, what value should be returned. The default value is true.

:filter_with - If set to :error, the query will raise an error on a match. If set to :filter the query will filter out unauthorized access. Valid values are :filter, :error The default value is :filter.

:validate? (boolean/0) - Whether or not to treat an invalid action as a non-allowed action. The default value is false.

:reuse_values? (boolean/0) - Whether or not loaded data like aggregates, calculations and relationships should be checked in memory if possible, instead of querying. No effect if pre_flight? is false. The default value is false.

:pre_flight? (boolean/0) - Whether or not this is a pre_flight check (which may perform optimized in-memory checks) or the final proper check. The default value is true.

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol. Will overwrite any actor, tenant or context provided. See Ash.Context for more.

:run_queries? (boolean/0) - Whether or not to run queries. If set to true, :maybe will not be returned. The default value is true.

:data - The record or records specifically attempting to be acted upon.

:tenant (value that implements the Ash.ToTenant protocol) - The tenant to use for authorization

:alter_source? (boolean/0) - If set to true, the source being authorized is returned so it can be run. The default value is false.

:base_query (term/0) - A base query on which to apply an generated filters

:no_check? (boolean/0) - Whether or not authorization must pass at the strict/filter step, or if post-checks are allowed to be run The default value is false.

:on_must_pass_strict_check (term/0) - Override the value returned when no_check? is true but a check must be run.

:atomic_changeset (term/0) - A base query on which to apply an generated filters

:return_forbidden_error? (boolean/0) - Whether or not to return a forbidden error in cases of not being authorized. The default value is false.

:log? (boolean/0) - Whether or not to log the authorization result. The default value is false.

changeset_to_create(name, params_or_opts \\ %{}, opts \\ [])

Returns the changeset corresponding to the action.

Options

:tracer (one or a list of module that adopts Ash.Tracer) - A tracer that implements the Ash.Tracer behaviour. See that module for more.

:authorize? - If an actor option is provided (even if it is nil), authorization happens automatically. If not, this flag can be used to authorize with no user.

:tenant (value that implements the Ash.ToTenant protocol) - A tenant to set on the query or changeset

:actor (term/0) - If an actor is provided, it will be used in conjunction with the authorizers of a resource to authorize access

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol, for passing around actor/tenant/context in a single value. See Ash.Scope.ToOpts for more.

:changeset (term/0) - A changeset to seed the action with.

changeset_to_update_statistics(record, params_or_opts \\ %{}, opts \\ [])

Returns the changeset corresponding to the action.

Options

:tracer (one or a list of module that adopts Ash.Tracer) - A tracer that implements the Ash.Tracer behaviour. See that module for more.

:authorize? - If an actor option is provided (even if it is nil), authorization happens automatically. If not, this flag can be used to authorize with no user.

:tenant (value that implements the Ash.ToTenant protocol) - A tenant to set on the query or changeset

:actor (term/0) - If an actor is provided, it will be used in conjunction with the authorizers of a resource to authorize access

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol, for passing around actor/tenant/context in a single value. See Ash.Scope.ToOpts for more.

create(name, params \\ nil, opts \\ nil)

Calls the create action on EhsEnforcement.Enforcement.Offender.

Arguments

name

Inputs

total_cases

county - County from EA structured address (e.g., 'SOUTH YORKSHIRE')

country

total_notices

company_registration_number - Companies House registration number (e.g., '04622955')

sic_code

main_activity

total_fines

business_type

first_seen_date

postcode

agencies

industry

local_authority

industry_sectors - EA detailed industry sectors (e.g., ['Manufacturing - General Engineering'])

town - Town from EA structured address (e.g., 'BARNSLEY')

last_seen_date

address

Options

:upsert? (boolean/0) - If a conflict is found based on the primary key, the record is updated in the database (requires upsert support) The default value is false.

:return_skipped_upsert? (boolean/0) - If true, and a record was not upserted because its filter prevented the upsert, the original record (which was not upserted) will be returned. The default value is false.

:upsert_identity (atom/0) - The identity to use when detecting conflicts for upsert?, e.g. upsert_identity: :full_name. By default, the primary key is used. Has no effect if upsert?: true is not provided

:upsert_fields - The fields to upsert. If not set, the action's upsert_fields is used, and if that is not set, then any fields not being set to defaults are written.

:upsert_condition (term/0) - An expression to check if the record should be updated when there's a conflict.

:timeout (timeout/0) - A positive integer, or :infinity. If none is provided, the timeout configured on the domain is used.

:tracer (one or a list of module that adopts Ash.Tracer) - A tracer that implements the Ash.Tracer behaviour. See that module for more.

:action (term/0) - The action to use, either an Action struct or the name of the action

:authorize? - If an actor option is provided (even if it is nil), authorization happens automatically. If not, this flag can be used to authorize with no user.

:context (map/0) - Context to set on the query, changeset, or input

:tenant (value that implements the Ash.ToTenant protocol) - A tenant to set on the query or changeset

:actor (term/0) - If an actor is provided, it will be used in conjunction with the authorizers of a resource to authorize access

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol, for passing around actor/tenant/context in a single value. See Ash.Scope.ToOpts for more.

:return_notifications? (boolean/0) - Use this if you're running ash actions in your own transaction and you want to manually handle sending notifications.
If a transaction is ongoing, and this is false, notifications will be discarded, otherwise the return value is {:ok, result, notifications} (or {:ok, notifications})
To send notifications later, use Ash.Notifier.notify(notifications). It sends any notifications that can be sent, and returns the rest. The default value is false.

:rollback_on_error? (boolean/0) - Whether or not to rollback the transaction on error, if the resource is in a transaction.
If the action has transaction? false this option has no effect. If an error is returned from the data layer and the resource is in a transaction, the transaction is always rolled back, regardless. The default value is true.

:notification_metadata (term/0) - Metadata to be merged into the metadata field for all notifications sent from this operation. The default value is %{}.

:skip_unknown_inputs - A list of inputs that, if provided, will be ignored if they are not recognized by the action. Use :* to indicate all unknown keys.

:load (term/0) - A load statement to add onto the changeset

:changeset (term/0) - A changeset to seed the action with.

:bulk_options (keyword/0) - Options passed to Ash.bulk_create, if a list or stream of inputs is provided.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:after_action (function of arity 2) - An after_action hook to be added to each processed changeset

:read_action (atom/0) - The action to use when building the read query.

:assume_casted? (boolean/0) - Whether or not to cast attributes and arguments as input. This is an optimization for cases where the input is already casted and/or not in need of casting The default value is false.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:authorize_query_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_changeset_with - If set to :error, instead of filtering unauthorized changes, unauthorized changes will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. The default value is :filter.

:private_arguments (map/0) - Private argument values to set on each changeset before validations and changes are run. The default value is %{}.

:sorted? (boolean/0) - Whether or not to sort results by their input position, in cases where return_records?: true was provided. The default value is false.

:return_records? (boolean/0) - Whether or not to return all of the records that were inserted. Defaults to false to account for large inserts. The default value is false.

:return_errors? (boolean/0) - Whether to return all errors that occur during the operation. Defaults to the value of :bulk_actions_default_to_errors? in your config, or false if not set. Returning all errors may be expensive for large inserts. The default value is false.

:batch_size (pos_integer/0) - The number of records to include in each batch. Defaults to the default_limit or max_page_size of the action, or 100.

:return_stream? (boolean/0) - If set to true, instead of an Ash.BulkResult, a mixed stream is returned. Potential elements: {:notification, notification} - if return_notifications? is set to true {:ok, record} - if return_records? is set to true {:error, error} - an error that occurred. May be changeset or an individual error. The default value is false.

:return_nothing? (boolean/0) - Mutes warnings about returning nothing. Only relevant if return_stream? is set to true and all other return_*? options are set to false. The default value is false.

:stop_on_error? (boolean/0) - If true, the first encountered error will stop the action and be returned. Otherwise, errors will be skipped. The default value is false.

:notify? (boolean/0) - Whether or not to generate any notifications. If this is set to true then the data layer must return the results from each batch. This may be intensive for large bulk actions. Notifications will be automatically sent unless return_notifications? is set to true. The default value is false.

:transaction - Whether or not to wrap the entire execution in a transaction, each batch, or not at all. Keep in mind: before_transaction and after_transaction hooks attached to changesets will have to be run inside the transaction if you choose transaction: :all. The default value is :batch.

:max_concurrency (non_neg_integer/0) - If set to a value greater than 0, up to that many tasks will be started to run batches asynchronously The default value is 0.

:private_arguments (map/0) - Private argument values to set before validations and changes. The default value is %{}.

create!(name, params \\ nil, opts \\ nil)

Calls the create action on EhsEnforcement.Enforcement.Offender. Raises any errors instead of returning them

Arguments

name

Inputs

total_cases

county - County from EA structured address (e.g., 'SOUTH YORKSHIRE')

country

total_notices

company_registration_number - Companies House registration number (e.g., '04622955')

sic_code

main_activity

total_fines

business_type

first_seen_date

postcode

agencies

industry

local_authority

industry_sectors - EA detailed industry sectors (e.g., ['Manufacturing - General Engineering'])

town - Town from EA structured address (e.g., 'BARNSLEY')

last_seen_date

address

Options

:upsert? (boolean/0) - If a conflict is found based on the primary key, the record is updated in the database (requires upsert support) The default value is false.

:return_skipped_upsert? (boolean/0) - If true, and a record was not upserted because its filter prevented the upsert, the original record (which was not upserted) will be returned. The default value is false.

:upsert_identity (atom/0) - The identity to use when detecting conflicts for upsert?, e.g. upsert_identity: :full_name. By default, the primary key is used. Has no effect if upsert?: true is not provided

:upsert_fields - The fields to upsert. If not set, the action's upsert_fields is used, and if that is not set, then any fields not being set to defaults are written.

:upsert_condition (term/0) - An expression to check if the record should be updated when there's a conflict.

:timeout (timeout/0) - A positive integer, or :infinity. If none is provided, the timeout configured on the domain is used.

:tracer (one or a list of module that adopts Ash.Tracer) - A tracer that implements the Ash.Tracer behaviour. See that module for more.

:action (term/0) - The action to use, either an Action struct or the name of the action

:authorize? - If an actor option is provided (even if it is nil), authorization happens automatically. If not, this flag can be used to authorize with no user.

:context (map/0) - Context to set on the query, changeset, or input

:tenant (value that implements the Ash.ToTenant protocol) - A tenant to set on the query or changeset

:actor (term/0) - If an actor is provided, it will be used in conjunction with the authorizers of a resource to authorize access

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol, for passing around actor/tenant/context in a single value. See Ash.Scope.ToOpts for more.

:return_notifications? (boolean/0) - Use this if you're running ash actions in your own transaction and you want to manually handle sending notifications.
If a transaction is ongoing, and this is false, notifications will be discarded, otherwise the return value is {:ok, result, notifications} (or {:ok, notifications})
To send notifications later, use Ash.Notifier.notify(notifications). It sends any notifications that can be sent, and returns the rest. The default value is false.

:rollback_on_error? (boolean/0) - Whether or not to rollback the transaction on error, if the resource is in a transaction.
If the action has transaction? false this option has no effect. If an error is returned from the data layer and the resource is in a transaction, the transaction is always rolled back, regardless. The default value is true.

:notification_metadata (term/0) - Metadata to be merged into the metadata field for all notifications sent from this operation. The default value is %{}.

:skip_unknown_inputs - A list of inputs that, if provided, will be ignored if they are not recognized by the action. Use :* to indicate all unknown keys.

:load (term/0) - A load statement to add onto the changeset

:changeset (term/0) - A changeset to seed the action with.

:bulk_options (keyword/0) - Options passed to Ash.bulk_create, if a list or stream of inputs is provided.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:after_action (function of arity 2) - An after_action hook to be added to each processed changeset

:read_action (atom/0) - The action to use when building the read query.

:assume_casted? (boolean/0) - Whether or not to cast attributes and arguments as input. This is an optimization for cases where the input is already casted and/or not in need of casting The default value is false.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:authorize_query_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_changeset_with - If set to :error, instead of filtering unauthorized changes, unauthorized changes will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. The default value is :filter.

:private_arguments (map/0) - Private argument values to set on each changeset before validations and changes are run. The default value is %{}.

:sorted? (boolean/0) - Whether or not to sort results by their input position, in cases where return_records?: true was provided. The default value is false.

:return_records? (boolean/0) - Whether or not to return all of the records that were inserted. Defaults to false to account for large inserts. The default value is false.

:return_errors? (boolean/0) - Whether to return all errors that occur during the operation. Defaults to the value of :bulk_actions_default_to_errors? in your config, or false if not set. Returning all errors may be expensive for large inserts. The default value is false.

:batch_size (pos_integer/0) - The number of records to include in each batch. Defaults to the default_limit or max_page_size of the action, or 100.

:return_stream? (boolean/0) - If set to true, instead of an Ash.BulkResult, a mixed stream is returned. Potential elements: {:notification, notification} - if return_notifications? is set to true {:ok, record} - if return_records? is set to true {:error, error} - an error that occurred. May be changeset or an individual error. The default value is false.

:return_nothing? (boolean/0) - Mutes warnings about returning nothing. Only relevant if return_stream? is set to true and all other return_*? options are set to false. The default value is false.

:stop_on_error? (boolean/0) - If true, the first encountered error will stop the action and be returned. Otherwise, errors will be skipped. The default value is false.

:notify? (boolean/0) - Whether or not to generate any notifications. If this is set to true then the data layer must return the results from each batch. This may be intensive for large bulk actions. Notifications will be automatically sent unless return_notifications? is set to true. The default value is false.

:transaction - Whether or not to wrap the entire execution in a transaction, each batch, or not at all. Keep in mind: before_transaction and after_transaction hooks attached to changesets will have to be run inside the transaction if you choose transaction: :all. The default value is :batch.

:max_concurrency (non_neg_integer/0) - If set to a value greater than 0, up to that many tasks will be started to run batches asynchronously The default value is 0.

:private_arguments (map/0) - Private argument values to set before validations and changes. The default value is %{}.

default_short_name()

find_or_create_offender(attrs)

Finds or creates an offender with deduplication logic.

This function handles finding existing offenders by name and postcode, with fuzzy matching fallback to prevent duplicates.

input(opts)

@spec input(values :: map() | Keyword.t()) :: map() | no_return()

Validates that the keys in the provided input are valid for at least one action on the resource.

Raises a KeyError error at compile time if not. This exists because generally a struct should only ever be created by Ash as a result of a successful action. You should not be creating records manually in code, e.g %MyResource{value: 1, value: 2}. Generally that is fine, but often with embedded resources it is nice to be able to validate the keys that are being provided, e.g

Resource
|> Ash.Changeset.for_create(:create, %{embedded: EmbeddedResource.input(foo: 1, bar: 2)})
|> Ash.create()

input(opts, action)

@spec input(values :: map() | Keyword.t(), action :: atom()) :: map() | no_return()

Same as input/1, except restricts the keys to values accepted by the action provided.

normalize_company_name(name)

Normalize company names to standard format for matching.

primary_key_matches?(left, right)

query_to_search(query, params_or_opts \\ %{}, opts \\ [])

Returns the query corresponding to the action.

Options

:tracer (one or a list of module that adopts Ash.Tracer) - A tracer that implements the Ash.Tracer behaviour. See that module for more.

:authorize? - If an actor option is provided (even if it is nil), authorization happens automatically. If not, this flag can be used to authorize with no user.

:tenant (value that implements the Ash.ToTenant protocol) - A tenant to set on the query or changeset

:actor (term/0) - If an actor is provided, it will be used in conjunction with the authorizers of a resource to authorize access

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol, for passing around actor/tenant/context in a single value. See Ash.Scope.ToOpts for more.

:query - A query to seed the action with.

search(query, params \\ nil, opts \\ nil)

Calls the search action on EhsEnforcement.Enforcement.Offender.

Arguments

query

Options

:page - Pagination options, see Ash.read/2 for more.

:load (term/0) - A load statement to add onto the query

:max_concurrency (non_neg_integer/0) - The maximum number of processes allowed to be started for parallel loading of relationships and calculations. Defaults to System.schedulers_online() * 2

:lock (term/0) - A lock statement to add onto the query

:return_query? (boolean/0) - If true, the query that was ultimately used is returned as a third tuple element.
The query goes through many potential changes during a request, potentially adding authorization filters, or replacing relationships for other data layers with their corresponding ids. This option can be used to get the true query that was sent to the data layer. The default value is false.

:skip_unknown_inputs - A list of inputs that, if provided, will be ignored if they are not recognized by the action. Use :* to indicate all unknown keys.

:reuse_values? (boolean/0) - Whether calculations are allowed to reuse values that have already been loaded, or must refetch them from the data layer. The default value is false.

:strict? (boolean/0) - If set to true, only specified attributes will be loaded when passing a list of fields to fetch on a relationship, which allows for more optimized data-fetching.
See Ash.Query.load/2. The default value is false.

:authorize_with - If set to :error, instead of applying authorization filters as a filter, any records not matching the authorization filter will cause an error to be returned. Valid values are :filter, :error The default value is :filter.

:timeout (timeout/0) - A positive integer, or :infinity. If none is provided, the timeout configured on the domain is used.

:tracer (one or a list of module that adopts Ash.Tracer) - A tracer that implements the Ash.Tracer behaviour. See that module for more.

:action (term/0) - The action to use, either an Action struct or the name of the action

:authorize? - If an actor option is provided (even if it is nil), authorization happens automatically. If not, this flag can be used to authorize with no user.

:context (map/0) - Context to set on the query, changeset, or input

:tenant (value that implements the Ash.ToTenant protocol) - A tenant to set on the query or changeset

:actor (term/0) - If an actor is provided, it will be used in conjunction with the authorizers of a resource to authorize access

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol, for passing around actor/tenant/context in a single value. See Ash.Scope.ToOpts for more.

:query - A query to seed the action with.

:not_found_error? (boolean/0) - Whether or not to return or raise a NotFound error or to return nil when a get? action/interface is called.

:stream? (boolean/0) - If true, a stream of the results will be returned The default value is false.

:stream_options (keyword/0) - Options passed to Ash.stream!, if stream?: true is given

:batch_size (integer/0) - How many records to request in each query run. Defaults to the pagination limits on the resource, or 250.

:allow_stream_with - The 'worst' strategy allowed to be used to fetch records. See Ash.stream!/2 docs for more. The default value is :keyset.

:stream_with - The specific strategy to use to fetch records. See Ash.stream!/2 docs for more.

search!(query, params \\ nil, opts \\ nil)

Calls the search action on EhsEnforcement.Enforcement.Offender. Raises any errors instead of returning them

Arguments

query

Options

:page - Pagination options, see Ash.read/2 for more.

:load (term/0) - A load statement to add onto the query

:max_concurrency (non_neg_integer/0) - The maximum number of processes allowed to be started for parallel loading of relationships and calculations. Defaults to System.schedulers_online() * 2

:lock (term/0) - A lock statement to add onto the query

:return_query? (boolean/0) - If true, the query that was ultimately used is returned as a third tuple element.
The query goes through many potential changes during a request, potentially adding authorization filters, or replacing relationships for other data layers with their corresponding ids. This option can be used to get the true query that was sent to the data layer. The default value is false.

:skip_unknown_inputs - A list of inputs that, if provided, will be ignored if they are not recognized by the action. Use :* to indicate all unknown keys.

:reuse_values? (boolean/0) - Whether calculations are allowed to reuse values that have already been loaded, or must refetch them from the data layer. The default value is false.

:strict? (boolean/0) - If set to true, only specified attributes will be loaded when passing a list of fields to fetch on a relationship, which allows for more optimized data-fetching.
See Ash.Query.load/2. The default value is false.

:authorize_with - If set to :error, instead of applying authorization filters as a filter, any records not matching the authorization filter will cause an error to be returned. Valid values are :filter, :error The default value is :filter.

:timeout (timeout/0) - A positive integer, or :infinity. If none is provided, the timeout configured on the domain is used.

:tracer (one or a list of module that adopts Ash.Tracer) - A tracer that implements the Ash.Tracer behaviour. See that module for more.

:action (term/0) - The action to use, either an Action struct or the name of the action

:authorize? - If an actor option is provided (even if it is nil), authorization happens automatically. If not, this flag can be used to authorize with no user.

:context (map/0) - Context to set on the query, changeset, or input

:tenant (value that implements the Ash.ToTenant protocol) - A tenant to set on the query or changeset

:actor (term/0) - If an actor is provided, it will be used in conjunction with the authorizers of a resource to authorize access

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol, for passing around actor/tenant/context in a single value. See Ash.Scope.ToOpts for more.

:query - A query to seed the action with.

:not_found_error? (boolean/0) - Whether or not to return or raise a NotFound error or to return nil when a get? action/interface is called.

:stream? (boolean/0) - If true, a stream of the results will be returned The default value is false.

:stream_options (keyword/0) - Options passed to Ash.stream!, if stream?: true is given

:batch_size (integer/0) - How many records to request in each query run. Defaults to the pagination limits on the resource, or 250.

:allow_stream_with - The 'worst' strategy allowed to be used to fetch records. See Ash.stream!/2 docs for more. The default value is :keyset.

:stream_with - The specific strategy to use to fetch records. See Ash.stream!/2 docs for more.

update_statistics(record, params \\ nil, opts \\ nil)

Calls the update_statistics action on EhsEnforcement.Enforcement.Offender.

Inputs

fine_amount

Options

:params (map/0) - Parameters to supply, ignored if the input is a changeset, only used when an identifier is given.

:atomic_upgrade? (boolean/0) - If true the action will be done atomically if it can (and is configured to do so), ignoring the in memory transformations and validations. You should not generally need to disable this. The default value is true.

:timeout (timeout/0) - A positive integer, or :infinity. If none is provided, the timeout configured on the domain is used.

:tracer (one or a list of module that adopts Ash.Tracer) - A tracer that implements the Ash.Tracer behaviour. See that module for more.

:action (term/0) - The action to use, either an Action struct or the name of the action

:authorize? - If an actor option is provided (even if it is nil), authorization happens automatically. If not, this flag can be used to authorize with no user.

:context (map/0) - Context to set on the query, changeset, or input

:tenant (value that implements the Ash.ToTenant protocol) - A tenant to set on the query or changeset

:actor (term/0) - If an actor is provided, it will be used in conjunction with the authorizers of a resource to authorize access

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol, for passing around actor/tenant/context in a single value. See Ash.Scope.ToOpts for more.

:return_notifications? (boolean/0) - Use this if you're running ash actions in your own transaction and you want to manually handle sending notifications.
If a transaction is ongoing, and this is false, notifications will be discarded, otherwise the return value is {:ok, result, notifications} (or {:ok, notifications})
To send notifications later, use Ash.Notifier.notify(notifications). It sends any notifications that can be sent, and returns the rest. The default value is false.

:rollback_on_error? (boolean/0) - Whether or not to rollback the transaction on error, if the resource is in a transaction.
If the action has transaction? false this option has no effect. If an error is returned from the data layer and the resource is in a transaction, the transaction is always rolled back, regardless. The default value is true.

:notification_metadata (term/0) - Metadata to be merged into the metadata field for all notifications sent from this operation. The default value is %{}.

:skip_unknown_inputs - A list of inputs that, if provided, will be ignored if they are not recognized by the action. Use :* to indicate all unknown keys.

:load (term/0) - A load statement to add onto the changeset

:bulk_options (keyword/0) - Options passed to Ash.bulk_update, if a query, list, or stream of inputs is provided.

:atomic_update (map/0) - A map of atomic updates to apply. See Ash.Changeset.atomic_update/3 for more.

:stream_batch_size (integer/0) - Batch size to use if provided a query and the query must be streamed

:allow_stream_with - The 'worst' strategy allowed to be used to fetch records if the :stream strategy is chosen. See the Ash.stream!/2 docs for more. The default value is :keyset.

:authorize_query? (boolean/0) - If a query is given, determines whether or not authorization is run on that query. The default value is true.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:filter (term/0) - A filter to apply to records. This is also applied to a stream of inputs.

:strategy - The strategy or strategies to enable. :stream is used in all cases if the data layer does not support atomics. The default value is [:atomic].

:allow_stream_with - The 'worst' strategy allowed to be used to fetch records. See Ash.stream!/2 docs for more. The default value is :keyset.

:stream_with - The specific strategy to use to fetch records. See Ash.stream!/2 docs for more.

:max_concurrency (non_neg_integer/0) - The maximum number of processes allowed to be started for parallel loading of relationships and calculations. Defaults to System.schedulers_online() * 2

:lock (term/0) - A lock statement to add onto the query

:return_query? (boolean/0) - If true, the query that was ultimately used is returned as a third tuple element. The query goes through many potential changes during a request, potentially adding authorization filters, or replacing relationships for other data layers with their corresponding ids. This option can be used to get the true query that was sent to the data layer. The default value is false.

:reuse_values? (boolean/0) - Whether calculations are allowed to reuse values that have already been loaded, or must refetch them from the data layer. The default value is false.

:strict? (boolean/0) - If set to true, only specified attributes will be loaded when passing a list of fields to fetch on a relationship, which allows for more optimized data-fetching. See Ash.Query.load/2. The default value is false.

:authorize_with - If set to :error, instead of applying authorization filters as a filter, any records not matching the authorization filter will cause an error to be returned. The default value is :filter.

:read_action (atom/0) - The action to use when building the read query.

:assume_casted? (boolean/0) - Whether or not to cast attributes and arguments as input. This is an optimization for cases where the input is already casted and/or not in need of casting The default value is false.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:authorize_query_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_changeset_with - If set to :error, instead of filtering unauthorized changes, unauthorized changes will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. The default value is :filter.

:private_arguments (map/0) - Private argument values to set on each changeset before validations and changes are run. The default value is %{}.

:sorted? (boolean/0) - Whether or not to sort results by their input position, in cases where return_records?: true was provided. The default value is false.

:return_records? (boolean/0) - Whether or not to return all of the records that were inserted. Defaults to false to account for large inserts. The default value is false.

:return_errors? (boolean/0) - Whether to return all errors that occur during the operation. Defaults to the value of :bulk_actions_default_to_errors? in your config, or false if not set. Returning all errors may be expensive for large inserts. The default value is false.

:batch_size (pos_integer/0) - The number of records to include in each batch. Defaults to the default_limit or max_page_size of the action, or 100.

:return_stream? (boolean/0) - If set to true, instead of an Ash.BulkResult, a mixed stream is returned. Potential elements: {:notification, notification} - if return_notifications? is set to true {:ok, record} - if return_records? is set to true {:error, error} - an error that occurred. May be changeset or an individual error. The default value is false.

:return_nothing? (boolean/0) - Mutes warnings about returning nothing. Only relevant if return_stream? is set to true and all other return_*? options are set to false. The default value is false.

:stop_on_error? (boolean/0) - If true, the first encountered error will stop the action and be returned. Otherwise, errors will be skipped. The default value is false.

:notify? (boolean/0) - Whether or not to generate any notifications. If this is set to true then the data layer must return the results from each batch. This may be intensive for large bulk actions. Notifications will be automatically sent unless return_notifications? is set to true. The default value is false.

:transaction - Whether or not to wrap the entire execution in a transaction, each batch, or not at all. Keep in mind: before_transaction and after_transaction hooks attached to changesets will have to be run inside the transaction if you choose transaction: :all. The default value is :batch.

:max_concurrency (non_neg_integer/0) - If set to a value greater than 0, up to that many tasks will be started to run batches asynchronously The default value is 0.

:private_arguments (map/0) - Private argument values to set before validations and changes. The default value is %{}.

update_statistics!(record, params \\ nil, opts \\ nil)

Calls the update_statistics action on EhsEnforcement.Enforcement.Offender. Raises any errors instead of returning them

Inputs

fine_amount

Options

:params (map/0) - Parameters to supply, ignored if the input is a changeset, only used when an identifier is given.

:atomic_upgrade? (boolean/0) - If true the action will be done atomically if it can (and is configured to do so), ignoring the in memory transformations and validations. You should not generally need to disable this. The default value is true.

:timeout (timeout/0) - A positive integer, or :infinity. If none is provided, the timeout configured on the domain is used.

:tracer (one or a list of module that adopts Ash.Tracer) - A tracer that implements the Ash.Tracer behaviour. See that module for more.

:action (term/0) - The action to use, either an Action struct or the name of the action

:authorize? - If an actor option is provided (even if it is nil), authorization happens automatically. If not, this flag can be used to authorize with no user.

:context (map/0) - Context to set on the query, changeset, or input

:tenant (value that implements the Ash.ToTenant protocol) - A tenant to set on the query or changeset

:actor (term/0) - If an actor is provided, it will be used in conjunction with the authorizers of a resource to authorize access

:scope (term/0) - A value that implements the Ash.Scope.ToOpts protocol, for passing around actor/tenant/context in a single value. See Ash.Scope.ToOpts for more.

:return_notifications? (boolean/0) - Use this if you're running ash actions in your own transaction and you want to manually handle sending notifications.
If a transaction is ongoing, and this is false, notifications will be discarded, otherwise the return value is {:ok, result, notifications} (or {:ok, notifications})
To send notifications later, use Ash.Notifier.notify(notifications). It sends any notifications that can be sent, and returns the rest. The default value is false.

:rollback_on_error? (boolean/0) - Whether or not to rollback the transaction on error, if the resource is in a transaction.
If the action has transaction? false this option has no effect. If an error is returned from the data layer and the resource is in a transaction, the transaction is always rolled back, regardless. The default value is true.

:notification_metadata (term/0) - Metadata to be merged into the metadata field for all notifications sent from this operation. The default value is %{}.

:skip_unknown_inputs - A list of inputs that, if provided, will be ignored if they are not recognized by the action. Use :* to indicate all unknown keys.

:load (term/0) - A load statement to add onto the changeset

:bulk_options (keyword/0) - Options passed to Ash.bulk_update, if a query, list, or stream of inputs is provided.

:atomic_update (map/0) - A map of atomic updates to apply. See Ash.Changeset.atomic_update/3 for more.

:stream_batch_size (integer/0) - Batch size to use if provided a query and the query must be streamed

:allow_stream_with - The 'worst' strategy allowed to be used to fetch records if the :stream strategy is chosen. See the Ash.stream!/2 docs for more. The default value is :keyset.

:authorize_query? (boolean/0) - If a query is given, determines whether or not authorization is run on that query. The default value is true.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:filter (term/0) - A filter to apply to records. This is also applied to a stream of inputs.

:strategy - The strategy or strategies to enable. :stream is used in all cases if the data layer does not support atomics. The default value is [:atomic].

:allow_stream_with - The 'worst' strategy allowed to be used to fetch records. See Ash.stream!/2 docs for more. The default value is :keyset.

:stream_with - The specific strategy to use to fetch records. See Ash.stream!/2 docs for more.

:max_concurrency (non_neg_integer/0) - The maximum number of processes allowed to be started for parallel loading of relationships and calculations. Defaults to System.schedulers_online() * 2

:lock (term/0) - A lock statement to add onto the query

:return_query? (boolean/0) - If true, the query that was ultimately used is returned as a third tuple element. The query goes through many potential changes during a request, potentially adding authorization filters, or replacing relationships for other data layers with their corresponding ids. This option can be used to get the true query that was sent to the data layer. The default value is false.

:reuse_values? (boolean/0) - Whether calculations are allowed to reuse values that have already been loaded, or must refetch them from the data layer. The default value is false.

:strict? (boolean/0) - If set to true, only specified attributes will be loaded when passing a list of fields to fetch on a relationship, which allows for more optimized data-fetching. See Ash.Query.load/2. The default value is false.

:authorize_with - If set to :error, instead of applying authorization filters as a filter, any records not matching the authorization filter will cause an error to be returned. The default value is :filter.

:read_action (atom/0) - The action to use when building the read query.

:assume_casted? (boolean/0) - Whether or not to cast attributes and arguments as input. This is an optimization for cases where the input is already casted and/or not in need of casting The default value is false.

:select (list of atom/0) - A select statement to apply to records. Ignored if return_records? is not true.

:authorize_query_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_changeset_with - If set to :error, instead of filtering unauthorized changes, unauthorized changes will raise an appropriate forbidden error. Uses authorize_with if not set.

:authorize_with - If set to :error, instead of filtering unauthorized query results, unauthorized query results will raise an appropriate forbidden error. The default value is :filter.

:private_arguments (map/0) - Private argument values to set on each changeset before validations and changes are run. The default value is %{}.

:sorted? (boolean/0) - Whether or not to sort results by their input position, in cases where return_records?: true was provided. The default value is false.

:return_records? (boolean/0) - Whether or not to return all of the records that were inserted. Defaults to false to account for large inserts. The default value is false.

:return_errors? (boolean/0) - Whether to return all errors that occur during the operation. Defaults to the value of :bulk_actions_default_to_errors? in your config, or false if not set. Returning all errors may be expensive for large inserts. The default value is false.

:batch_size (pos_integer/0) - The number of records to include in each batch. Defaults to the default_limit or max_page_size of the action, or 100.

:return_stream? (boolean/0) - If set to true, instead of an Ash.BulkResult, a mixed stream is returned. Potential elements: {:notification, notification} - if return_notifications? is set to true {:ok, record} - if return_records? is set to true {:error, error} - an error that occurred. May be changeset or an individual error. The default value is false.

:return_nothing? (boolean/0) - Mutes warnings about returning nothing. Only relevant if return_stream? is set to true and all other return_*? options are set to false. The default value is false.

:stop_on_error? (boolean/0) - If true, the first encountered error will stop the action and be returned. Otherwise, errors will be skipped. The default value is false.

:notify? (boolean/0) - Whether or not to generate any notifications. If this is set to true then the data layer must return the results from each batch. This may be intensive for large bulk actions. Notifications will be automatically sent unless return_notifications? is set to true. The default value is false.

:transaction - Whether or not to wrap the entire execution in a transaction, each batch, or not at all. Keep in mind: before_transaction and after_transaction hooks attached to changesets will have to be run inside the transaction if you choose transaction: :all. The default value is :batch.

:max_concurrency (non_neg_integer/0) - If set to a value greater than 0, up to that many tasks will be started to run batches asynchronously The default value is 0.

:private_arguments (map/0) - Private argument values to set before validations and changes. The default value is %{}.