EhsEnforcement.Integrations.Airtable.UkAirtable (ehs_enforcement v0.1.0)

Types

opts()

@type opts() :: %{
  base_id: String.t(),
  table: String.t(),
  view: String.t(),
  fields: [String.t()],
  formula: String.t()
}

Functions

enumerate_at_records(arg, func)

enumerate_at_records(records, opts, func)

get_legal_article_taxa_records(opts)

@spec get_legal_article_taxa_records(opts()) :: [
  EhsEnforcement.Legislation.LegalRegister.legal_register()
]

get_legal_interpretation_records(opts)

@spec get_legal_interpretation_records(opts()) :: list()

get_legal_register_records(opts)

@spec get_legal_register_records(opts()) :: [
  EhsEnforcement.Legislation.LegalRegister.legal_register()
]

get_records_from_at(opts)

@spec get_records_from_at({:params, opts()}) :: {:ok, list()} | {:error, term()}

@spec get_records_from_at(opts()) :: {:ok, list()} | {:error, term()}

make_records_into_legal_article_taxa_structs(records)

@spec make_records_into_legal_article_taxa_structs(list()) :: [
  EhsEnforcement.Legislation.Taxa.LATTaxa.legal_register()
]

make_records_into_legal_register_structs(records)

@spec make_records_into_legal_register_structs(list()) :: [
  EhsEnforcement.Legislation.LegalRegister.legal_register()
]

move_id_and_drop_createdtime_field(records)

@spec move_id_and_drop_createdtime_field(list()) :: list()

Receives the records returned from an Airtable GET request Returns a list of maps with :id in fields, :createdtime and :fields members removed

strip_id_and_createdtime_fields(records)

@spec strip_id_and_createdtime_fields(list()) :: list()

Receives the records returned from an Airtable GET request Returns a list of maps with :id, :createdtime and :fields members removed