EhsEnforcement.Legislation.Taxa.LATTaxa (ehs_enforcement v0.1.0)

Module to generate taxa classes for sub-sections / sub-articles High level

Legl.Countries.Uk.Article.Taxa.LATTaxa

workflow -

Loops across all records tagging Duty Actors. These are roles present in the script and this is a 'broad brush' trawl.

Uses those results to seed the next process which is to tag duty types and the related dutyholders.

The same duty type process then tags the remaining framework clauses like amendments and offences.

The last process sets POPIMAR tags for those records tagged with duties.

Types

legal_article_taxa()

@type legal_article_taxa() :: %EhsEnforcement.Legislation.Taxa.LATTaxa{
  Chapter: String.t(),
  "Duty Actor": list(),
  "Duty Actor Aggregate": list(),
  "Duty Actor Gvt": list(),
  "Duty Actor Gvt Aggregate": list(),
  "Duty Type": list(),
  "Duty Type Aggregate": list(),
  Dutyholder: list(),
  "Dutyholder Aggregate": list(),
  Heading: String.t(),
  ID: String.t(),
  Number: String.t(),
  POPIMAR: list(),
  "POPIMAR Aggregate": list(),
  Part: String.t(),
  Power_Holder: list(),
  Power_Holder_Aggregate: list(),
  Record_ID: String.t(),
  Record_Type: list(),
  Responsibility_Holder: list(),
  Responsibility_Holder_Aggregate: list(),
  Rights_Holder: list(),
  Rights_Holder_Aggregate: list(),
  "Section||Regulation": String.t(),
  Text: String.t(),
  Year: integer(),
  dutyholder_txt: String.t(),
  power_holder_txt: String.t(),
  regexes: list(),
  responsibility_holder_txt: String.t(),
  rights_holder_txt: String.t(),
  type_code: String.t()
}

opts()

@type opts() :: map()

taxa()

@type taxa() :: atom()

Functions

aggregate(records, arg)

@spec aggregate(struct(), {taxa(), taxa()}) :: {:ok, map()}

Function aggregates sub-section and sub-article duty type tag at the level of section/article.

source from :"Duty Actor", :Dutyholder, :"Duty Type", :POPIMAR

aggregate_keys(records)

@spec aggregate_keys(struct()) :: list()

aggregate_result(aggregates, aggregate_field, records)

@spec aggregate_result(list(), atom(), struct()) :: list()

aggregates(keys, source_field, records)

@spec aggregates(list(), atom(), struct()) :: list()

api_update_lat_taxa(opts \\ [])

Function to process models (taxa) for Legal Article Table records

Workflow is driven by the user options. Main workflow is 'Update'

Functions comprising the workflow are passed to update_lat_taxa to be run

api_update_lat_taxa_from_text(records, opts)

@spec api_update_lat_taxa_from_text(
  [
    %UK.Act{
      amendment: term(),
      changes: term(),
      chapter: term(),
      figure_counter: term(),
      flow: term(),
      heading: term(),
      heading?: term(),
      id: term(),
      max_amendments: term(),
      max_commencements: term(),
      max_editorials: term(),
      max_extents: term(),
      max_modifications: term(),
      name: term(),
      para: term(),
      part: term(),
      region: term(),
      section: term(),
      sub_para: term(),
      sub_section: term(),
      table_counter: term(),
      text: term(),
      type: term()
    }
  ]
  | [
      %UK.Regulation{
        amendment: term(),
        changes: term(),
        chapter: term(),
        figure_counter: term(),
        flow: term(),
        heading: term(),
        heading?: term(),
        id: term(),
        max_amendments: term(),
        max_commencements: term(),
        max_editorials: term(),
        max_extents: term(),
        max_modifications: term(),
        name: term(),
        para: term(),
        part: term(),
        region: term(),
        section: term(),
        sub_para: term(),
        sub_section: term(),
        table_counter: term(),
        text: term(),
        type: term()
      }
    ],
  map()
) :: [
  %EhsEnforcement.Legislation.Taxa.LATTaxa{
    Chapter: term(),
    "Duty Actor": term(),
    "Duty Actor Aggregate": term(),
    "Duty Actor Gvt": term(),
    "Duty Actor Gvt Aggregate": term(),
    "Duty Type": term(),
    "Duty Type Aggregate": term(),
    Dutyholder: term(),
    "Dutyholder Aggregate": term(),
    Heading: term(),
    ID: term(),
    Number: term(),
    POPIMAR: term(),
    "POPIMAR Aggregate": term(),
    Part: term(),
    Power_Holder: term(),
    Power_Holder_Aggregate: term(),
    Record_ID: term(),
    Record_Type: term(),
    Responsibility_Holder: term(),
    Responsibility_Holder_Aggregate: term(),
    Rights_Holder: term(),
    Rights_Holder_Aggregate: term(),
    "Section||Regulation": term(),
    Text: term(),
    Year: term(),
    dutyholder_txt: term(),
    power_holder_txt: term(),
    regexes: term(),
    responsibility_holder_txt: term(),
    rights_holder_txt: term(),
    type_code: term()
  }
]

Function to process models (taxa) for parsed .html from leg.gov.uk

Receives either a %UK.Act{} or %UK.Regulation{} struct and returns %Legl.Countries.Uk.Article.Taxa.LATTaxa{}

api_update_multi_lat_taxa(opts \\ [])

duty_actor_aggregate(records)

duty_actor_gvt_aggregate(records)

duty_type_aggregate(records)

dutyholder_aggregate(records)

filesave(records, path, opts)

get(opts)

patch(results, opts)

popimar_aggregate(records)

power_holder_aggregate(records)

pre_process(records, blacklist)

Function to remove terms and phrases from the text (section etc) that are not a dutyholder or duty but could be confused as such. The amended text is saved back into the Taxa struct as "aText"

responsibility_holder_aggregate(records)

rightsholder_aggregate(records)

save_to_lat(records, opts)