EhsEnforcement.Scraping.ScrapeSession (ehs_enforcement v0.1.0)

Ash resource for tracking HSE scraping sessions with real-time progress updates.

Types

t()

@type t() :: %EhsEnforcement.Scraping.ScrapeSession{
  __lateral_join_source__: term(),
  __meta__: term(),
  __metadata__: term(),
  __order__: term(),
  action_types: term(),
  agency: term(),
  aggregates: term(),
  calculations: term(),
  cases_created: term(),
  cases_created_current_page: term(),
  cases_exist_current_page: term(),
  cases_exist_total: term(),
  cases_found: term(),
  cases_processed: term(),
  cases_updated: term(),
  cases_updated_current_page: term(),
  current_page: term(),
  database: term(),
  date_from: term(),
  date_to: term(),
  end_page: term(),
  errors_count: term(),
  id: term(),
  inserted_at: term(),
  max_pages: term(),
  pages_processed: term(),
  session_id: term(),
  start_page: term(),
  status: term(),
  updated_at: term()
}

Functions

default_short_name()

input(opts)

@spec input(values :: map() | Keyword.t()) :: map() | no_return()

Validates that the keys in the provided input are valid for at least one action on the resource.

Raises a KeyError error at compile time if not. This exists because generally a struct should only ever be created by Ash as a result of a successful action. You should not be creating records manually in code, e.g %MyResource{value: 1, value: 2}. Generally that is fine, but often with embedded resources it is nice to be able to validate the keys that are being provided, e.g

Resource
|> Ash.Changeset.for_create(:create, %{embedded: EmbeddedResource.input(foo: 1, bar: 2)})
|> Ash.create()

input(opts, action)

@spec input(values :: map() | Keyword.t(), action :: atom()) :: map() | no_return()

Same as input/1, except restricts the keys to values accepted by the action provided.

primary_key_matches?(left, right)