EhsEnforcementWeb.CaseLive.CSVExport (ehs_enforcement v0.1.0)

Handles CSV export functionality for case data

Functions

export_case(case_id)

Export a single case to CSV format

export_cases(filters \\ %{}, sort_by \\ :offence_action_date, sort_dir \\ :desc)

Export cases to CSV format based on current filters

export_detailed_cases(filters \\ %{}, sort_by \\ :offence_action_date, sort_dir \\ :desc)

Export cases with detailed information including related notice data

generate_csv(cases)

Generate CSV content from list of cases

generate_detailed_csv(cases_with_notices)

Generate detailed CSV content from list of cases with notice data

generate_filename(export_type \\ :all, identifier \\ nil)

Generate filename for CSV export